Revert "made PostNote an abstract class with child classes TextPostNote and ImagePostNote"
