+++ src/main/java/entity/ImagePostNote.java
@@ -1,19 +1,15 @@
 package entity;
 
-import view.NotePanel;
-
+import interface_adapter.create_MindMap.SquarePanel;
 import javax.swing.*;
-import java.awt.Color;
-import java.awt.Graphics;
-import java.awt.Point;
-import java.awt.event.MouseEvent;
+import java.awt.*;
 
 public class ImagePostNote extends PostNote {
     private ImageIcon imageIcon;
 
     // Constructor for ImagePostNote
-    public ImagePostNote(int x, int y, NotePanel panel) {
-        super(x, y, 0, 0, Color.WHITE, panel);  // Start with 0 width and height, we'll set those later
+    public ImagePostNote(int x, int y, Color color, SquarePanel panel) {
+        super(x, y, 0, 0, color, panel);  // Start with 0 width and height, we'll set those later
     }
 
     // Image inside post it note
@@ -35,7 +31,8 @@ public class ImagePostNote extends PostNote {
     // Override the draw method to handle custom image drawing inside the Post-it note
     @Override
     public void draw(Graphics g) {
-        label.setLocation(x, y);
+        super.draw(g);
+
         if (imageIcon != null) {
             // Draw the image inside the Post-it note with updated size
             g.drawImage(imageIcon.getImage(), x, y, imageIcon.getIconWidth(), imageIcon.getIconHeight(), null);
@@ -44,32 +41,11 @@ public class ImagePostNote extends PostNote {
 
     @Override
     public void dragTo(Point point) {
-        x = point.x - relativePos.x;
-        y = point.y - relativePos.y - 50;
-        label.setLocation(x, y);
-        panel.repaint();
+        super.dragTo(point);
 
         // Adjust image position along with PostNote if needed
         if (imageIcon != null) {
             // Optionally adjust position of the image relative to the post-it note
         }
-    }
-
-    /**
-     * Displays a context menu with options to edit, connect, or delete the postit.
-     * Prevents deletion of the initial postit.
-     *
-     * @param e MouseEvent that triggered the menu
-     */
-    @Override
-    protected void showContextMenu(MouseEvent e) {
-        JPopupMenu menu = new JPopupMenu();
-        System.out.println("pop up menu");
-        // calls startEditing when that option is pressed.
-        JMenuItem editPostItItem = new JMenuItem("Edit Post-it note");
-        // editPostItItem.addActionListener(event -> startEditing());
-
-        menu.add(editPostItItem);
-        menu.show(label, e.getX(), e.getY());
     }
 }
