--- src/main/java/interface_adapter/change_color/ChangeColorPresenter.java
@@ -1,4 +1,0 @@
-package interface_adapter.change_color;
-
-public class ChangeColorPresenter {
-}
