+++ src/main/java/interface_adapter/create_MindMap/SquarePanel.java
@@ -1,8 +1,7 @@
-package view;
+package interface_adapter.create_MindMap;
 
 import entity.ImagePostNote;
 import entity.PostNote;
-import entity.TextPostNote;
 
 import javax.swing.*;
 import java.awt.*;
@@ -12,10 +11,10 @@ import java.util.ArrayList;
 /**
  * A child class that inherits JPanel, it will hold all the post-it notes and take care of adding/removing/editing
  */
-public class NotePanel extends JPanel {
+public class SquarePanel extends JPanel {
     private ArrayList<PostNote> postNotes; // List to hold all post-it notes on the panel
 
-    public NotePanel(ArrayList<PostNote> postNotes) {
+    public SquarePanel(ArrayList<PostNote> postNotes) {
         this.postNotes = postNotes;
         setLayout(null); // Disable layout manager for manual positioning
         setBackground(Color.WHITE); // Set background color of the panel
@@ -40,7 +39,7 @@ public class NotePanel extends JPanel {
         JPopupMenu menu = new JPopupMenu();
 
         JMenuItem createNoteItem = new JMenuItem("Create New Post-it Note");
-        createNoteItem.addActionListener(event -> addTextPostNote());
+        createNoteItem.addActionListener(event -> createPostNote());
 
         menu.add(createNoteItem);
         menu.show(this, point.x, point.y);
@@ -49,18 +48,18 @@ public class NotePanel extends JPanel {
     /**
      * Creates a new post-it note at the specified position.
      */
-    public void addTextPostNote() {
-        final TextPostNote textPostNote = new TextPostNote(100, 100, 100, 100, Color.ORANGE, this);
-        add(textPostNote.getLabel());
-        add(textPostNote.getTextField());
+    public void createPostNote() {
+        PostNote newPostNote = new PostNote(100, 100, 100, 100, Color.ORANGE, this);
+        add(newPostNote.getLabel());
+        add(newPostNote.getTextField());
         repaint();
-        postNotes.add(textPostNote);
+        postNotes.add(newPostNote);
     }
 
     /**
      * Creates a new ImagePostNote at the specified position.
      */
-    public void addImagePostNote(ImagePostNote imagePostNote) {
+    public void createPostNote(ImagePostNote imagePostNote) {
         add(imagePostNote.getLabel());
         repaint();
         postNotes.add(imagePostNote);
