--- src/main/java/use_case/change_color/ChangeColorInteractor.java
@@ -1,4 +1,0 @@
-package use_case.change_color;
-
-public class ChangeColorInteractor {
-}
