--- src/main/java/use_case/change_color/ChangeColorOutputBoundary.java
@@ -1,15 +1,0 @@
-package use_case.change_color;
-
-public interface ChangeColorOutputBoundary {
-    /**
-     * Prepares the success view for the Change Color Use Case.
-     * @param outputData the output data
-     */
-    void prepareSuccessView(ChangeColorOutputData outputData);
-
-    /**
-     * Prepares the failure view for the Change Color Use Case.
-     * @param errorMessage the explanation of the failure
-     */
-    void prepareFailView(String errorMessage);
-}
