--- src/main/java/use_case/change_color/ChangeColorOutputData.java
@@ -1,4 +1,0 @@
-package use_case.change_color;
-
-public class ChangeColorOutputData {
-}
