--- src/main/java/use_case/change_color/ChangeColorUserDataAccessInterface.java
@@ -1,11 +1,0 @@
-package use_case.change_color;
-
-import entity.PostNote;
-
-public interface ChangeColorUserDataAccessInterface {
-    /**
-     * Updates the system to record this user's password.
-     * @param postNote the postNote whose color is to be changed
-     */
-    void changeColor(PostNote postNote);
-}
